project/actuators/requirements.txt
paho-mqtt<2.0.0
bottle==0.13.2
requests==2.32.3